backend/requirements.txt
anyascii==0.3.3
asgiref==3.9.1
beautifulsoup4==4.13.4
black==25.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cssbeautifier==1.15.4
defusedxml==0.7.1
Django==5.2.5
django-ckeditor==6.7.3
django-cors-headers==4.7.0
django-filter==25.1
django-js-asset==3.1.2
django-modelcluster==6.4
django-nested-admin==4.1.3
django-permissionedforms==0.1
django-rest-framework==0.1.0
django-stubs-ext==5.2.2
django-taggit==6.1.0
django-tasks==0.8.1
django-treebeard==4.7.1
djangorestframework==3.16.1
djlint==1.36.4
draftjs_exporter==5.1.0
drf-nested-routers==0.94.2
EditorConfig==0.17.1
et_xmlfile==2.0.0
filetype==1.2.0
flake8==7.3.0
gunicorn==23.0.0
idna==3.10
jedi==0.17.2
jsbeautifier==1.15.4
json5==0.12.1
laces==0.1.2
mccabe==0.7.0
mypy==1.17.1
mypy_extensions==1.1.0
nodeenv==1.9.1
openpyxl==3.1.5
packaging==25.0
parso==0.7.1
pathspec==0.12.1
pillow==11.3.0
pillow_heif==1.1.0
platformdirs==4.3.8
pluggy==1.6.0
psycopg==3.2.9
pycodestyle==2.14.0
pyflakes==3.4.0
pyright==1.1.404
python-decouple==3.8
python-dotenv==1.1.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-monkey-business==1.1.0
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
telepath==0.3.1
tqdm==4.67.1
typing_extensions==4.14.1
ujson==5.11.0
urllib3==2.5.0
wagtail==7.1
Willow==1.11.0
